Feat: Catalog WIP - Add tabs and organize buttons
