+++ src/components/TopTabs/styles.module.css
@@ -1,0 +1,31 @@
+.tab {
+  text-decoration: none;
+  display: block;
+  padding: 15px 20px;
+  margin-top: 10px;
+  margin-right: 10px;
+  font-family: var(--comfortaa);
+  font-weight: bold;
+  border-top-right-radius: 10px;
+  border-top-left-radius: 10px;
+  outline: none;
+}
+
+.tab:active,
+.tab:visited,
+.tab:focus,
+.tab {
+  color: #000;
+}
+
+.tab.active {
+  background: linear-gradient(to bottom, var(--primary-card-bg) 20%, var(--off-white));
+  border: 0.5px solid var(--border-gray);
+  border-top-width: 2px;
+  border-bottom: none;
+}
+
+.tab:hover,
+.tab:focus-visible {
+  background-color: var(--off-white-hover);
+}
