+++ src/routes/Catalog/index.tsx
@@ -1,0 +1,3 @@
+export default function Catalog() {
+  return <h1>Catalog</h1>;
+}
