+++ src/styles/variables.css
@@ -37,6 +37,7 @@
   --primary-card-bg: #f5f6e6;
   --secondary-card-bg: #afd0d4;
   --off-white: #f3f3f3;
+  --off-white-hover: #f9f9f9;
   --quiz-card-bg-1: #009688;
   --quiz-card-bg-2: #2196f3;
   --quiz-card-bg-3: #e91e63;
